### 2017-07-21 #325 @whedon
[PRE REVIEW]: pyrgg: Python Random Graph Generator
**Submitting author:** @sepandhaghighi (<a href="http://orcid.org/0000-0001-9450-2375">Sepand Haghighi</a>)
**Repository:** <a href="https://github.com/sepandhaghighi/pyrgg" target ="_blank">https://github.com/sepandhaghighi/pyrgg</a>
**Version:** v0.1
