### 2017-07-25 d671a94
docs : README file updated
Former-commit-id: d6dcdf4595a9a809294f83bac51ed0bcb9a6c8d0
