### 2017-07-26 edb50b8
docs : FullRun video added
Former-commit-id: 78b1ab70f0d4dbe571fdb0bdb6e6233b16ddf8d3
