### 2017-07-26 8f44233
docs : paper file updated
Former-commit-id: e400658f18de2dae9c532894b5a0ce1cf3edae0b
