### 2017-07-26 a71625d
docs : README file updated
Former-commit-id: 7323c72dc7868161200808264260de91a8951840
