### 2017-07-29 sepandhaghighi/pyrgg#2 @hausen
JOSS submission acceptance blockers
* Missing statement of need: the software generates random graphs, but the user can only choose the number of vertices, min and max weight for the edges, and a range for a random number of edges. I would expect a useful random graph generator to provide some more options, like connectivity, min, max and avg degree.

* Example usage: the README.md states that the output _can_ be used as input to other programs, but have the files generated by this program ever been used as an input to some other algorithm? how?

* Tests: no tests provided

* Other:
  - there is a "gpu" tag in paper.md, why?
  - ideally, large binary files as the ones in the [datasets](https://github.com/sepandhaghighi/pyrgg/tree/master/datasets) dir should not belong in a git repository. See: http://blog.deveo.com/storing-large-binary-files-in-git-repositories/
