### 2017-07-29 #331 @hausen
Issues: https://github.com/sepandhaghighi/pyrgg/issues/2
