+++ .git-rewrite/t/README.md
@@ -1,0 +1,47 @@
+
+# Python Random Graph Generator							
+
+
+
+## Overview
+Simple random graph generator based on <a href ="http://www.cse.psu.edu/~kxm85/software/GTgraph/index.html">GTgraph</a>
+
+
+
+## Installation ##
+```python
+
+python setup.py install
+ 
+```
+
+
+## Usage ##
+- Just enter number of vertices
+
+
+## Issues & Bug Reports			
+
+Just fill an issue and describe it. I'll check it ASAP!							
+or send an email to [sepand@qpage.ir](mailto:sepand@qpage.ir "sepand@qpage.ir"). 
+
+## TODO		
+
+- [x] Small Graph 
+- [x] Weighted Graph
+- [x] Directed Graph
+- [ ] Huge Graph Generation
+- [ ] Multithreading
+
+
+
+
+## Contribution			
+
+You can fork the repository, improve or fix some part of it and then send the pull requests back if you want to see them here. I really appreciate that. ❤️			
+
+Remember to write a few tests for your code before sending pull requests.  			
+
+## License
+
+<a href="https://github.com/sepandhaghighi/qpage/blob/master/LICENSE"><img src="https://img.shields.io/github/license/mashape/apistatus.svg"/></a>   
