### 2017-07-29 6093b28
fix : doctest example added
