### 2017-07-29 9d1271a
fix : test function deleted from pyrgg.py
