### 2017-07-29 faa70a5
fix : systemrandom convert to psuedo random
