### 2017-07-29 f158a6e
fix : git-rewrite folder deleted
