+++ .git-rewrite/t/setup.py
@@ -1,15 +1,0 @@
-from distutils.core import setup
-setup(
-  name = 'Python Graph Generator',
-  packages = ['pyrgg'],
-  version = '0.1',
-  description = 'Simple Random Graph Generator',
-  long_description="",
-  author = 'Sepand Haghighi',
-  author_email = 'sepand@qpage.ir',
-  url = 'https://github.com/sepandhaghighi/pyrgg',
-  download_url = 'https://github.com/sepandhaghighi/pyrgg/tarball/v0.1',
-  keywords = ['random', 'graph', 'python3','python','generator'],
-  classifiers = [],
-  license='MIT',
-)
