### 2017-07-29 28e645e
docs : gpu tag deleted from paper.md
