### 2017-07-29 sepandhaghighi/pyrgg#2 @sepandhaghighi
Hi @hausen

Thanks for review ;-)

1- About statement of need  :  this software is a quick&easy random graph generator and generate input file for other graph processing applications and one of it goals is small input. But your are right, with few more options it can be more useful.

2- About example of usage : yes some of graph processing frameworks like medusa  , cusha and ... use this graph format (DIMACS) , and first pyrgg written to generate synthetic inputs for these frameworks benchmarking against native code in a paper. Then released as a stand-alone software

3- About test : I added [test cases](https://github.com/sepandhaghighi/pyrgg#test) for functions.

4- I deleted gpu tag from [paper.md](https://github.com/sepandhaghighi/pyrgg/blob/master/paper/paper.md)

5- Thanks for your hint about large files and repo ;-)
I deleted datasets folder and all of its traces from github with BFG Repo-Cleaner and now size of repo is about 6MB (prev about 90MB!!)

6- Software references are in [paper.bib](https://github.com/sepandhaghighi/pyrgg/blob/master/paper/paper.bib)

Best Regards
Sepand Haghighi
