+++ .travis.yml
@@ -1,0 +1,12 @@
+sudo: true
+
+language: python
+
+python:
+  - 3.5
+
+install:
+  - python3 setup install
+
+script:
+  - python3 -m pyrgg test
