### 2017-07-29 689d21c
.travis.yml file updated
