+++ README.md
@@ -55,7 +55,7 @@ this software generate input file for graph processing applications.
 
 `python3 -m pyrgg test` or `python -m pyrgg test`
 			
-
+This test run automatically in each commit ;-) <a href="https://travis-ci.org/sepandhaghighi/pyrgg"><img src="https://travis-ci.org/sepandhaghighi/pyrgg.svg?branch=master"></a>				
 <div align="center">
 
 <img src="http://www.shaghighi.ir/pyrgg/images/test.gif" alt="pyrgg test" title="pyrgg test">
