### 2017-08-01 17d983d
fix : minor edit in tests
