+++ .travis.yml
@@ -10,3 +10,4 @@ install:
 
 script:
   - python3 -m pyrgg test
+  - cat testfile.gr
