+++ appveyor.yml
@@ -1,0 +1,21 @@
+build: false
+
+environment:
+  matrix:
+    - PYTHON: "C:\\Python33"
+      PYTHON_VERSION: "3.3.5"
+      PYTHON_ARCH: "32"
+
+    - PYTHON: "C:\\Python34"
+      PYTHON_VERSION: "3.4.1"
+      PYTHON_ARCH: "32"
+
+
+init:
+  - "ECHO %PYTHON% %PYTHON_VERSION% %PYTHON_ARCH%"
+
+install:
+  - "python setup.py install"
+
+test_script:
+  - "python -m pyrgg test"
