### 2017-08-02 1883f77
fix : minor edit in appveyor config file
