+++ appveyor.yml
@@ -3,12 +3,12 @@ build: false
 environment:
   matrix:
     - PYTHON: "C:\\Python33"
-      PYTHON_VERSION: "3.3.5"
+      PYTHON_VERSION: "3.5.1"
       PYTHON_ARCH: "32"
 
     - PYTHON: "C:\\Python34"
-      PYTHON_VERSION: "3.4.1"
-      PYTHON_ARCH: "32"
+      PYTHON_VERSION: "3.5.2"
+      PYTHON_ARCH: "64"
 
 
 init:
