### 2017-08-02 259122e
minor edit in appveyor config file
