+++ appveyor.yml
@@ -2,12 +2,9 @@ build: false
 
 environment:
   matrix:
-    - PYTHON: "C:\\Python33"
-      PYTHON_VERSION: "3.5.1"
-      PYTHON_ARCH: "32"
 
-    - PYTHON: "C:\\Python34"
-      PYTHON_VERSION: "3.5.2"
+    - PYTHON: "C:\\Python35"
+      PYTHON_VERSION: "3.5"
       PYTHON_ARCH: "64"
 
 
