### 2017-08-07 8614659
fix : minor edit in appveyor config for python  version
