### 2017-08-07 ad72c4d
docs : appveyor badge added to README file
