+++ README.md
@@ -55,7 +55,12 @@ this software generate input file for graph processing applications.
 
 `python3 -m pyrgg test` or `python -m pyrgg test`
 			
-This test run automatically in each commit ;-) <a href="https://travis-ci.org/sepandhaghighi/pyrgg"><img src="https://travis-ci.org/sepandhaghighi/pyrgg.svg?branch=master"></a>				
+This test run automatically in each commit ;-) 						
+
+Linux : <a href="https://travis-ci.org/sepandhaghighi/pyrgg"><img src="https://travis-ci.org/sepandhaghighi/pyrgg.svg?branch=master"></a>							
+
+Windows : <a href="https://ci.appveyor.com/project/sepandhaghighi/pyrgg"><img src="https://ci.appveyor.com/api/projects/status/j9g3xikmh1elti13?svg=true"></a>						
+			
 <div align="center">
 
 <img src="http://www.shaghighi.ir/pyrgg/images/test.gif" alt="pyrgg test" title="pyrgg test">
