### 2017-08-07 sepandhaghighi/pyrgg#2 @sepandhaghighi
Hi @hausen

[Appveyor](https://ci.appveyor.com/project/sepandhaghighi/pyrgg) and [TravisCI](https://travis-ci.org/sepandhaghighi/pyrgg) added for automated tests.

Best Regards
Sepand Haghighi
