### 2017-08-07 #331 @danielskatz
@hausen - looking at the comments in https://github.com/sepandhaghighi/pyrgg/issues/2, it seems like is ready for your re-review
