### 2017-08-14 331b8b9
docs : paper file updated
