+++ paper/paper.bib
@@ -17,3 +17,15 @@
  address = {New York, NY, USA},
  keywords = {GPGPU, GPU programming, graph processing, runtime framework},
 }
+
+@incollection{Chakrabarti2004,
+  doi = {10.1137/1.9781611972740.43},
+  url = {https://doi.org/10.1137/1.9781611972740.43},
+  year  = {2004},
+  month = {apr},
+  publisher = {Society for Industrial and Applied Mathematics},
+  pages = {442--446},
+  author = {Deepayan Chakrabarti and Yiping Zhan and Christos Faloutsos},
+  title = {R-{MAT}: A Recursive Model for Graph Mining},
+  booktitle = {Proceedings of the 2004 {SIAM} International Conference on Data Mining}
+}
