### 2017-08-16 sepandhaghighi/pyrgg#2 @hausen
hi @sepandhaghighi ,

1) I don't want to sound too harsh, but your program can be summarized as:
```
    given: n (number of vertices), min_num_edges, max_num_edges, min_weight, max_weight
    num_edges = random(min_num_edges, max_num_edges)
    for 1...num_edges
        print edge from random(1,n) to random(1,n) with weight random(min_weight, max_weight)
```
As you said, it is a quick and easy random graph generator. But a statement of need must explain why someone would pick your random graph generator instead of coding their own or using another one with more features. For instance: your code will likely generate graphs with isolated vertices, but for most uses, isolated vertices can (or must) be eliminated.

2) The question is "have the files generated by this program ever been used as an input to some other algorithm? how?" Can you please provide a real-world example of your code being used as an input to those graph processing frameworks? Is it being used in research? Or was it used before?

3) The tests are not comprehensive; generation of files in the DIMACS format is not covered.
