### 2017-08-16 #331 @hausen
@danielskatz , was busy for a couple of days, sorry. I've just re-reviewed it. There are still some issues.
