### 2017-08-18 0330e81
docs : new test case added
