### 2017-08-18 c0e6543
fix : file_maker renamed to dimacs
