+++ pyrgg/test.py
@@ -35,7 +35,7 @@ This function get a string as input if input is one digit add a zero
 >>> edge_gen(20,0,100,2,10,2)
 [{1: [18, 15, 17, 7, 21, 6, 5, 20, 1], 2: [2, 7, 20], 3: [11, 19, 17, 21, 16, 3, 14, 9, 8], 4: [1, 19, 4, 13, 7, 16], 5: [14, 19, 7, 9, 3, 11, 4, 8], 6: [1, 15, 6], 7: [7, 17, 5, 21, 4, 14, 1, 19, 6, 20], 8: [2, 7, 9], 9: [10, 3, 19, 8, 20, 12, 15], 10: [19, 13, 21, 10, 20, 7, 14, 4, 2], 11: [8, 13, 14, 16, 17, 3], 12: [16, 21, 20, 9, 7], 13: [3, 14], 14: [2, 6, 12, 19, 3], 15: [15, 17, 5, 2], 16: [12, 10, 1, 21, 16, 5, 3, 18, 2], 17: [11, 3, 16, 14], 18: [19, 20, 13, 3, 21, 9, 11, 15, 18], 19: [17, 10, 3, 1, 4, 20, 16, 11, 15, 8], 20: [9, 21, 7, 13, 19, 5, 12, 3]}, {1: [99, 57, 75, 23, 78, 12, 11, 50, 67], 2: [4, 30, 3], 3: [56, 25, 29, 37, 0, 58, 70, 40, 65], 4: [8, 98, 51, 8, 6, 48], 5: [9, 80, 99, 43, 39, 1, 17, 90], 6: [7, 62, 87], 7: [57, 24, 53, 49, 50, 27, 34, 38, 50, 82], 8: [18, 56, 1], 9: [49, 9, 81, 1, 47, 79, 16], 10: [17, 23, 19, 29, 31, 20, 6, 13, 65], 11: [94, 32, 76, 37, 22, 16], 12: [71, 78, 9, 27, 95], 13: [34, 57], 14: [5, 36, 67, 16, 46], 15: [42, 74, 75, 2], 16: [89, 4, 76, 9, 8, 9, 57, 47, 94], 17: [45, 87, 9, 3], 18: [1, 84, 48, 11, 14, 53, 49, 59, 10], 19: [3, 76, 61, 29, 63, 84, 32, 84, 63, 41], 20: [25, 55, 27, 28, 40, 63, 5, 35]}, 129]
 >>> random.seed(2)
->>> file_maker("testfile", 0, 200, 10, 0, 2, 0)
+>>> dimacs_maker("testfile", 0, 200, 10, 0, 2, 0)
 7
 >>> file=open("testfile.gr","r")
 >>> print(file.read())
@@ -56,7 +56,7 @@ a 8 8 -97
 a 9 9 143
 <BLANKLINE>
 >>> random.seed(4)
->>> file_maker("testfile2",0,50,30,0,4,0)
+>>> dimacs_maker("testfile2",0,50,30,0,4,0)
 41
 >>> file=open("testfile2.gr","r")
 >>> print(file.read())
@@ -111,7 +111,7 @@ a 26 19 28
 a 27 14 7
 <BLANKLINE>
 >>> random.seed(20)
->>> file_maker("testfile3",10,30,100,0,4,2)
+>>> dimacs_maker("testfile3",10,30,100,0,4,2)
 191
 >>> file=open("testfile3.gr","r")
 >>> print(file.read())
