### 2017-08-18 fe81079
docs : minor edit in paper file
