### 2017-08-18 sepandhaghighi/pyrgg#2 @sepandhaghighi
Hi @hausen

Thanks for  re-review :-)

1- From one sight your opinion is right, this program is not complex but we can see lots of useful packages with small functions and also same package in different programming languages!
As I mentioned before this package is an easy and quick graph generator and the main aspect of pyrgg is programmability that generate synthetic random graphs easily.
I'm a computer architecture researcher and work on machine learning and deep learning algorithms implementation on GPU(mainly  graph processing algorithms), for one of our research in lab we needed a graph generator  to create synthetic graphs in different sizes to benchmark our project, I searched for a package but there is no useful package in python that  generate DIMACS (need in project) format graphs so I write this project and share it as a stand-alone open source package for a researcher like me in the future.
I recently added JSON format and also have plan to develop this package and add ram management - optimum multi thread system to generate giant graphs,Erdős–Rényi model and ...
You talked about isolated vertices, you are right, in most of cases these vertices must be eliminated. But one part of our research is about a algorithm that need these isolated vertices as cases, so I didn't delete this option.

2- This package can be used for any research program that need DIMACS graph format, and answer of your question  is yes, pyrgg used as a generator in 3 GPU/FPGA based graph processing algorithms and these papers will be  published in future.

3- I add more tests for DIMACS and JSON file formats. [Travis](https://travis-ci.org/sepandhaghighi/pyrgg) , [Appveyor](https://ci.appveyor.com/project/sepandhaghighi/pyrgg)

Best Regards
Sepand Haghighi
