### 2017-08-18 93c9493
fix : minor edit in travis config file
