+++ .travis.yml
@@ -1,8 +1,5 @@
 sudo: true
 
-os:
-  - linux
-  - osx
 language: python
 
 python:
