### 2017-08-18 0e1f2b1
docs : minor edit in paper
