### 2017-08-18 e168363
docs : JSON sample links added to README file
