### 2017-08-18 519411f
docs : minor edit in README file
