### 2017-08-19 #331 @hausen
The author [states](https://github.com/sepandhaghighi/pyrgg/issues/2#issuecomment-323446098) that the software was created to fill a specific need of his lab and that it is being used to generate samples for benchmarking some undisclosed algorithms, the results of which will be released _in the future_.

Therefore, "statement of need," "example usage" and "references" are still missing.

Moreover, IMHO, the software is too simple to be of general use as is.

If the author disagrees, he should disclose more details of what the software is being used for, maybe by putting a preprint of the benchmarking results on Arxiv and showing how they were obtained.
