### 2017-08-19 #331 @sepandhaghighi
Hi

I tell a fact, I have a research project that need graphs ( different sizes )
in DIMACS/JSON format, and my framework written in python, what can I do ?  there is no python package that  generate graphs in this format. I think this fact is enough for statement of need.(Is not specific need, most of hardware graph processing framework in hardware level use DIMACS/JSON format)

I'm not owner of research projects mentioned before, so unfortunately is not possible to share and publish any parts of them.

Your opinion is respected but I think you take a little hard, you tell that pyrgg is too simple, I don't know whats the problem ? There are  lots of software and packages that are not complex but there are useful in research projects.

most of converters and generators in research projects focused on formatting and automation so there are not complex in algorithm,  like [country converter coco](http://joss.theoj.org/papers/af694f2e5994b8aacbad119c4005e113) that accepted in joss, if we want to summarized its like a simple table  but it is very useful.

I think at least there is no problem with my [reference](https://github.com/sepandhaghighi/pyrgg/blob/master/paper/paper.bib), both of them have DOI .

Best Regards
Sepand Haghighi
