### 2017-08-22 #331 @danielskatz
@hausen - Regarding references, please be aware this is a question about the references in the paper itself: Are there appropriate references (citations), and do they have DOIs if there are DOIs for them?
