+++ README.md
@@ -16,9 +16,8 @@
 </div>				
 				
 ## Overview
-pyrgg is a synthetic random graph generator based on <a href ="http://www.diag.uniroma1.it/challenge9/format.shtml">DIMACS</a>/JSON format
-this software generate input file for graph processing applications.  
 
+pyrgg is an easy to use synthetic random graph generator based on <a href ="http://www.diag.uniroma1.it/challenge9/format.shtml">DIMACS</a>/JSON format written in python. 
 
 
 ## Installation		
@@ -117,9 +116,11 @@ or send an email to [sepand@qpage.ir](mailto:sepand@qpage.ir "sepand@qpage.ir").
 - [Sample 2-JSON](https://www.dropbox.com/s/f6kljlch7p2rfhy/1000.json?dl=0) (1000 Vertices , 105KB)
 
 
-## Application			
-This software can be used for generation of input files in parallel sparse graph applications benchmark like :			
+## Application
 
+pyrgg is an easy to use synthetic random graph generator based on <a href ="http://www.diag.uniroma1.it/challenge9/format.shtml">DIMACS</a>/JSON format written in python. 
+pyrgg has this ability to generate graphs in different sizes and designed to provide input files for test, benchmark and performance analysis of graph processing frameworks(some of them mentioned here) , but it can be used in any other research program that include graphs in DIMACS/JSON format.
+					
 
 - [Medusa](https://github.com/JianlongZhong/Medusa "Medusa") 
 - [Totem](https://github.com/netsyslab/Totem "Totem")
