### 2017-08-22 82ac2a1
docs : paper updated
