+++ README.md
@@ -119,7 +119,7 @@ or send an email to [sepand@qpage.ir](mailto:sepand@qpage.ir "sepand@qpage.ir").
 ## Application
 
 pyrgg is an easy to use synthetic random graph generator based on <a href ="http://www.diag.uniroma1.it/challenge9/format.shtml">DIMACS</a>/JSON format written in python. 
-pyrgg has this ability to generate graphs in different sizes and designed to provide input files for test, benchmark and performance analysis of graph processing frameworks(some of them mentioned here) , but it can be used in any other research program that include graphs in DIMACS/JSON format.
+pyrgg has the ability to generate graphs in different sizes and designed to provide input files for test, benchmark and performance analysis of graph processing frameworks(some of them mentioned here) , but it can be used in any other research program that include graphs in DIMACS/JSON format.
 					
 
 - [Medusa](https://github.com/JianlongZhong/Medusa "Medusa") 
