### 2017-08-23 cdf2d33
docs : TODO list updated
