+++ pyrgg/test.py
@@ -355,5 +355,262 @@ a 100 38 10
 '74'
 >>> testfile_3["graph"]["edges"][1]["weight"]
 '15'
+>>> random.seed(2)
+>>> csv_maker("testfile", 0, 200, 10, 0, 2, 0)
+7
+>>> file=open("testfile.csv","r")
+>>> print(file.read())
+4,3,-64
+5,4,148
+5,11,110
+6,7,-139
+6,8,-9
+8,8,-97
+9,9,143
+<BLANKLINE>
+>>> random.seed(4)
+>>> csv_maker("testfile2",0,50,30,0,4,0)
+41
+>>> file=open("testfile2.csv","r")
+>>> print(file.read())
+1,10,46
+2,16,5
+2,3,25
+2,18,-48
+4,17,-17
+5,27,16
+6,31,41
+7,26,-12
+7,6,-18
+8,3,-42
+8,13,11
+9,16,-5
+10,27,0
+10,30,-36
+10,23,-48
+10,17,26
+11,20,-27
+11,15,14
+11,10,-2
+14,21,-33
+14,18,-44
+14,11,43
+15,3,-12
+16,9,22
+16,14,-40
+16,18,20
+19,9,7
+20,30,18
+20,15,2
+21,23,18
+21,24,-1
+22,10,-9
+22,25,-39
+24,20,28
+24,10,16
+25,20,21
+25,19,23
+25,2,-10
+26,26,-18
+26,19,28
+27,14,7
+<BLANKLINE>
+>>> random.seed(20)
+>>> csv_maker("testfile3",10,30,100,0,4,2)
+191
+>>> file=open("testfile3.csv","r")
+>>> print(file.read())
+1,34,30
+3,74,15
+3,4,23
+4,10,13
+4,17,20
+4,61,28
+5,53,16
+5,26,20
+5,81,20
+6,55,12
+6,81,26
+7,52,12
+7,27,28
+7,32,11
+8,13,12
+9,33,19
+10,89,18
+10,21,29
+13,91,17
+13,29,28
+14,8,10
+15,39,18
+15,18,30
+15,43,29
+15,93,24
+16,82,18
+17,56,15
+17,22,22
+17,91,13
+17,89,14
+18,55,21
+18,29,13
+18,78,11
+18,67,16
+19,23,15
+19,11,13
+19,101,19
+21,87,22
+21,59,26
+23,16,24
+24,43,10
+24,10,28
+24,96,13
+24,12,20
+25,83,12
+25,85,19
+28,29,26
+29,90,10
+29,21,10
+29,86,26
+29,18,19
+30,54,18
+32,13,16
+32,31,16
+32,29,13
+33,17,27
+33,50,14
+35,63,10
+35,23,16
+35,82,19
+36,13,19
+36,27,21
+36,80,19
+37,87,13
+39,8,22
+39,48,19
+40,60,13
+40,9,28
+40,61,16
+42,9,27
+42,80,20
+42,92,10
+42,23,18
+43,78,13
+43,1,11
+44,96,16
+45,52,29
+45,11,28
+45,47,19
+45,57,11
+46,80,22
+47,51,22
+47,33,10
+47,100,24
+47,12,28
+48,41,17
+48,100,16
+48,38,24
+49,49,16
+49,80,11
+50,18,12
+50,96,21
+53,31,27
+53,19,19
+53,100,10
+54,88,21
+55,68,14
+55,3,17
+56,68,24
+56,32,22
+56,50,27
+57,65,29
+57,84,19
+58,98,12
+58,84,25
+58,35,21
+58,95,14
+59,85,17
+59,63,14
+59,46,20
+60,19,30
+60,96,27
[…]
