### 2017-08-23 234577a
docs : README file todo list updated with csv output format
