### 2017-08-23 10f58df
docs : paper and readme file updated
