### 2017-08-23 12fb177
docs : minor edit in README file
