### 2017-08-23 2d0b0f8
docs: README and Paper file updated
