### 2017-08-23 2e3e0f7
docs : minor edit in paper
