### 2017-08-23 4b496a8
docs : minor edit in paper
