### 2017-08-23 b0af580
fix :minor edit in paper spacing
