+++ build/lib/pyrgg/__init__.py
@@ -1,1 +1,0 @@
-from .pyrgg import *
