+++ build/lib/pyrgg/pyrgg.py
@@ -1,330 +1,0 @@
-import random
-import os
-import datetime
-import sys
-import doctest
-
-
-Source_dir=os.getcwd()
-#random_system=random.SystemRandom()
-random_system=random
-def logger(vertices_number,edge_number,file_name,elapsed_time):
-    '''
-    This function save generated graphs log
-    :param vertices_number: nubmer of vertices
-    :type vertices_number:int
-    :param edge_number: Number of edges
-    :type edge_number:int;
-    :param file_name:  File Name
-    :type file_name:str
-    :param elapsed_time: Elapsed Time
-     :type elapsed_time : str
-    :return:  None
-    '''
-    try:
-        file = open("logfile.log", "a")
-        file.write(str(datetime.datetime.now())+"\n")
-        file.write("Filename : "+file_name+"\n")
-        file.write("Vertices : "+str(vertices_number)+"\n")
-        file.write("Edges : "+str(edge_number)+"\n")
-        file.write("Elapsed Time : " + str(elapsed_time) + "\n")
-        file.write("-------------------------------\n")
-        file.close()
-    except Exception:
-        if file.closed==False:
-            file.close()
-        print("Error In Logger")
-
-def zero_insert(input_string):
-    '''
-    This function get a string as input if input is one digit add a zero
-    :param input_string: input digit az string
-    :type input_string:str
-    :return: modified output as str
-    >>> zero_insert("22")
-    >>> '22'
-    '''
-    if len(input_string)==1:
-        return "0"+input_string
-    return input_string
-
-def time_convert(input_string):
-    '''
-    This function convert input_string from sec to DD,HH,MM,SS Format
-    :param input_string: input time string  in sec
-    :type input_string:str
-    :return: converted time as string
-    '''
-    input_sec=float(input_string)
-    input_minute=input_sec//60
-    input_sec=int(input_sec-input_minute*60)
-    input_hour=input_minute//60
-    input_minute=int(input_minute-input_hour*60)
-    input_day=int(input_hour//24)
-    input_hour=int(input_hour-input_day*24)
-    return zero_insert(str(input_day))+" days, "+zero_insert(str(input_hour))+" hour, "+zero_insert(str(input_minute))+" minutes, "+zero_insert(str(input_sec))+" seconds"
-
-
-def get_input():
-    '''
-    This function get input from user and return as dictionary
-    :return: inputs as dictionary
-    '''
-    try:
-        file_name=input("File Name : ")
-        if file_name+".gr" in os.listdir():
-            raise Exception("There is file with this name")
-        vertices=int(input("Vertices Number : "))
-        max_weight=int(input("Max Weight : "))
-        min_weight = int(input("Min Weight : "))
-        min_edge=int(input("Min Edge Number :"))
-        min_edge=max(0,min_edge)
-        max_edge=int(input("Max Edge Number :"))
-        max_edge=min(max_edge,vertices)
-        sign_flag=int(input("Signed[1] or Unsigned[2]"))
-        output_format=int(input("DIMACS(.gr)[1] | JSON(.json)[2] | CSV(.csv)[3] | WEL(.wel)[4]"))
-        if sign_flag not in [1,2]:
-            sign_flag=2
-        if output_format not in [1,2,3,4]:
-            output_format=1
-        return {"file_name":file_name,"vertices":vertices,"max_weight":max_weight,"min_weight":min_weight,"min_edge":min_edge,"max_edge":max_edge,"sign":sign_flag,"output_format":output_format}
-    except Exception as e:
-        print(e)
-        sys.exit()
-
-def sign_gen():
-    '''
-    This function return random sign
-    :return: 1 or -1
-    '''
-    flag=random_system.randint(0,1)
[…]
