+++ .travis.yml
@@ -6,6 +6,7 @@ python:
   - 3.5
 
 install:
+  - pip3 install -r requirements.txt
   - python3 setup.py install
 
 script:
