+++ requirements.txt
@@ -1,3 +1,3 @@
-yaml
+pyyaml
 
 
