### 2017-08-25 1ebe2d4
appveyor config file fixed
