### 2017-08-25 17dafc6
appveyor config file fixed
