### 2017-08-25 21f5790
appveyor config file fixed
