+++ appveyor.yml
@@ -17,7 +17,7 @@ init:
   - "ECHO %PYTHON% %PYTHON_VERSION% %PYTHON_ARCH%"
 
 install:
-  - "%PYTHON%/pip3.exe -r requirements.txt"
+  - "%PYTHON%/Scripts/pip.exe -r requirements.txt"
   - "%PYTHON%/python.exe setup.py install"
 
 test_script:
