### 2017-08-25 7eab6c0
appveyor config file fixed
