### 2017-08-26 f088dfb
Exceptions test added
