+++ pyrgg/__main__.py
@@ -23,8 +23,11 @@ def run():
         json_to_yaml(file_name)
     elif input_dict["output_format"]==5 :
         edge_number = wel_maker(file_name, min_weight, max_weight, vertices_number, min_edge, max_edge, sign)
-    else:
+    elif input_dict["output_format"]==6:
         edge_number = lp_maker(file_name, min_weight, max_weight, vertices_number, min_edge, max_edge, sign)
+    else:
+        edge_number = json_maker(file_name, min_weight, max_weight, vertices_number, min_edge, max_edge, sign)
+        json_to_pickle(file_name)
     print("Generating . . . ")
     second_time = time.perf_counter()
     elapsed_time = second_time - first_time
