+++ pyrgg/test.py
@@ -341,6 +341,7 @@ a 100 38 10
 >>> file=open("testfile.json","r")
 >>> import json
 >>> import yaml
+>>> import pickle
 >>> testfile_1=json.load(file)
 >>> testfile_1["graph"]["nodes"][1]
 {'id': '2'}
@@ -358,6 +359,14 @@ a 100 38 10
 >>> testfile_1_yaml["graph"]["edges"][1]["target"]
 '4'
 >>> testfile_1_yaml["graph"]["edges"][1]["weight"]
+'148'
+>>> json_to_pickle("testfile")
+>>> testfile_1_p=pickle.load( open( "testfile.p", "rb" ) )
+>>> testfile_1_p["graph"]["edges"][1]["source"]
+'5'
+>>> testfile_1_p["graph"]["edges"][1]["target"]
+'4'
+>>> testfile_1_p["graph"]["edges"][1]["weight"]
 '148'
 >>> random.seed(4)
 >>> json_maker("testfile2",0,50,30,0,4,0)
@@ -382,6 +391,14 @@ a 100 38 10
 >>> testfile_2_yaml["graph"]["edges"][1]["target"]
 '16'
 >>> testfile_2_yaml["graph"]["edges"][1]["weight"]
+'5'
+>>> json_to_pickle("testfile2")
+>>> testfile_2_p=pickle.load( open( "testfile2.p", "rb" ) )
+>>> testfile_2_p["graph"]["edges"][1]["source"]
+'2'
+>>> testfile_2_p["graph"]["edges"][1]["target"]
+'16'
+>>> testfile_2_p["graph"]["edges"][1]["weight"]
 '5'
 >>> random.seed(20)
 >>> json_maker("testfile3",10,30,100,0,4,2)
@@ -415,6 +432,14 @@ UnboundLocalError: local variable 'yaml_file' referenced before assignment
 Traceback (most recent call last):
         ...
 TypeError: json_maker() missing 1 required positional argument: 'sign'
+>>> json_to_pickle("testfile3")
+>>> testfile_3_p=pickle.load( open( "testfile3.p", "rb" ) )
+>>> testfile_3_p["graph"]["edges"][1]["source"]
+'3'
+>>> testfile_3_p["graph"]["edges"][1]["target"]
+'74'
+>>> testfile_3_p["graph"]["edges"][1]["weight"]
+'15'
 >>> random.seed(2)
 >>> csv_maker("testfile", 0, 200, 10, 0, 2, 0)
 7
