### 2017-08-27 272d651
fix : minor edit in pyrgg unused codes
