### 2017-08-27 6a920a6
coverall commands added
