+++ pyrgg/test.py
@@ -5,6 +5,9 @@ This function get a string as input if input is one digit add a zero
 :type input_string:str
 :return: modified output as str
 >>> from pyrgg import *
+>>> import coverage
+>>> cov = coverage.Coverage()
+>>> cov.start()
 >>> zero_insert("22")
 '22'
 >>> zero_insert("320")
@@ -1062,5 +1065,7 @@ edge(26,26,-18).
 edge(26,19,28).
 edge(27,14,7).
 <BLANKLINE>
+>>> cov.stop()
+>>> cov.save()
 
 '''
