+++ requirements.txt
@@ -1,3 +1,4 @@
 pyyaml
+python-coveralls
 
 
