+++ requirements.txt
@@ -1,4 +1,4 @@
 pyyaml
-python-coveralls
+codecov
 
 
