+++ .travis.yml
@@ -12,4 +12,4 @@ install:
 script:
   - python3 -m pyrgg test
 after_success:
-  - coveralls
+  - codecov
