### 2017-08-27 4f1af3f
fix : new testcase added
