### 2017-08-27 65dd9c9
fix : minor edit in testcases
