+++ pyrgg/test.py
@@ -60,10 +60,6 @@ TypeError: edge_gen() missing 1 required positional argument: 'sign'
 >>> random.seed(2)
 >>> dimacs_maker("testfile", 0, 200, 10, 0, 2, 0)
 7
->>> dimacs_maker("testfile", 0, 200,0, 2, 0)
-Traceback (most recent call last):
-        ...
-TypeError: dimacs_maker() missing 1 required positional argument: 'sign'
 >>> file=open("testfile.gr","r")
 >>> print(file.read())
 c FILE                  :testfile.gr
@@ -433,9 +429,9 @@ TypeError: dimacs_maker() missing 1 required positional argument: 'sign'
 >>> testfile_3_yaml["graph"]["edges"][1]["weight"]
 '15'
 >>> json_to_yaml("testfile24")
-Traceback (most recent call last):
-        ...
-UnboundLocalError: local variable 'yaml_file' referenced before assignment
+[Error] Bad Input File
+>>> json_to_pickle("testfile24")
+[Error] Bad Input File
 >>> json_maker("testfile", 0, 200, 10, 0, 0)
 Traceback (most recent call last):
         ...
