+++ pyrgg/pyrgg.py
@@ -32,9 +32,7 @@ def logger(vertices_number,edge_number,file_name,elapsed_time):
         file.write("-------------------------------\n")
         file.close()
     except Exception:
-        if file.closed==False:
-            file.close()
-        print("Error In Logger")
+        print("[Error] Logger Faild!")
 
 def zero_insert(input_string):
     '''
@@ -90,7 +88,7 @@ def get_input():
             output_format=1
         return {"file_name":file_name,"vertices":vertices,"max_weight":max_weight,"min_weight":min_weight,"min_edge":min_edge,"max_edge":max_edge,"sign":sign_flag,"output_format":output_format}
     except Exception as e:
-        print(e)
+        print("[Error] Bad Input!")
         sys.exit()
 
 def sign_gen():
@@ -197,24 +195,17 @@ def dimacs_maker(file_name,min_range,max_range,vertices,min_edge,max_edge,sign):
     :type sign: int
     :return: edge_number
     '''
-    try:
-        file=open(file_name+".gr","w")
-        dicts=edge_gen(vertices,min_range,max_range,min_edge,max_edge,sign)
-        edge_dic=dicts[0]
-        weight_dic=dicts[1]
-        edge_number=dicts[2]
-        file_init(file,file_name,min_range,max_range,vertices,edge_number,min_edge,max_edge)
-        for i in edge_dic.keys():
-            for j,value in enumerate(edge_dic[i]):
-                file.write("a "+str(i)+" "+str(value)+" "+str(weight_dic[i][j])+"\n")
-        file.close()
-        return edge_number
-    except Exception:
-        print("Error In File Creation")
-        if file.closed==False:
-            file.close()
-        os.remove(os.path.join(Source_dir,file_name)+".gr")
-        sys.exit()
+    file=open(file_name+".gr","w")
+    dicts=edge_gen(vertices,min_range,max_range,min_edge,max_edge,sign)
+    edge_dic=dicts[0]
+    weight_dic=dicts[1]
+    edge_number=dicts[2]
+    file_init(file,file_name,min_range,max_range,vertices,edge_number,min_edge,max_edge)
+    for i in edge_dic.keys():
+        for j,value in enumerate(edge_dic[i]):
+            file.write("a "+str(i)+" "+str(value)+" "+str(weight_dic[i][j])+"\n")
+    file.close()
+    return edge_number
 
 def json_maker(file_name,min_range,max_range,vertices,min_edge,max_edge,sign):
     '''
@@ -231,31 +222,24 @@ def json_maker(file_name,min_range,max_range,vertices,min_edge,max_edge,sign):
     :type sign: int
     :return: edge_number
     '''
-    try:
-        file = open(file_name + ".json", "w")
-        dicts = edge_gen(vertices, min_range, max_range, min_edge, max_edge, sign)
-        edge_dic = dicts[0]
-        weight_dic = dicts[1]
-        edge_number = dicts[2]
-        nodes='\t\t\t"nodes":[\n'
-        edges='\t\t\t"edges":[\n'
-        for i in edge_dic.keys():
-            nodes=nodes+'\t\t\t{\n\t\t\t\t'+'"id": '+'"'+str(i)+'"\n\t\t\t},\n'
-            for j,value in enumerate(edge_dic[i]):
-                edges=edges+'\t\t\t{\n\t\t\t\t"source": '+'"'+str(i)+'",\n\t\t\t\t'+'"target": '+'"'+str(value)+'",\n\t\t\t\t'+'"weight": '+'"'+str(weight_dic[i][j])+'"\n\t\t\t},\n'
-        nodes=nodes[:-2]+"\n\t\t],\n"
-        edges=edges[:-2]+"\n\t\t]\n\t}\n}"
[…]
