### 2017-08-27 5c105f2
fix : some test in coverage.xml
