+++ .travis.yml
@@ -7,11 +7,11 @@ python:
 
 install:
   - pip3 install -r requirements.txt
+  - python3 setup.py install
 
 script:
-  - cd pyrgg
-  - python3 test.py
-  - cd ..
-  - python3 setup.py install
+  - python3 -m pyrgg test
+  - coverage xml
+  - cat coverage.xml
 after_success:
   - codecov
