### 2017-08-27 a3ba9e6
fix : some test in coverage.xml
