### 2017-08-27 79ba54d
fix : some test in coverage.xml
