### 2017-08-27 f0e6e07
docs : coverage badge added to README
