### 2017-08-27 6bf24ad
fix : testcase added for get_input function
