### 2017-08-28 #331 @danielskatz
@hausen - For the references item in the checklist:

Are there appropriate references (citations), and do they have DOIs if there are DOIs for them?

If so, please check off the references item in the checklist.  If not, please let us know what you think is needed.
