### 2017-08-28 #331 @danielskatz
@sepandhaghighi - please note some inputs from you are needed.  As stated above:

Please discuss in general, in your documentation, what you expect others to do with the software, and how this will support their research. They can be hypothetical.

Note this is intended to answer: "Do the authors clearly state what problems the software is designed to solve and who the target audience is?"

And you do need to provide some examples of usage.
