### 2017-08-28 0e4be72
docs : example of usage added to README file
