### 2017-08-28 237bed1
docs : TGF sample file links added
