### 2017-08-28 586a909
docs : minor edit in documentation syntax
