+++ js/script.js
@@ -1,0 +1,21 @@
+
+function redirect(flag){
+    switch(flag){
+        case 1:
+            window.open("https://saythanks.io/to/sepandhaghighi");
+            break;
+        case 2:
+            swal("Bitocin Wallet","1XGr9qbZjBpUQJJSB6WtgBQbDTgrhPLPA");
+            break;
+        case 3:
+            window.open("https://beerpay.io/sepandhaghighi/lightsout");
+            break;
+        case 4:
+            window.open("https://www.payping.ir/sepandhaghighi");
+            break;
+        default:
+            window.open("https://saythanks.io/to/sepandhaghighi");
+            
+    }
+    
+}
