+++ pyrgg/__main__.py
@@ -25,9 +25,11 @@ def run():
         edge_number = wel_maker(file_name, min_weight, max_weight, vertices_number, min_edge, max_edge, sign)
     elif input_dict["output_format"]==6:
         edge_number = lp_maker(file_name, min_weight, max_weight, vertices_number, min_edge, max_edge, sign)
-    else:
+    elif input_dict["output_format"] == 7:
         edge_number = json_maker(file_name, min_weight, max_weight, vertices_number, min_edge, max_edge, sign)
         json_to_pickle(file_name)
+    else:
+        edge_number=dl_maker(file_name, min_weight, max_weight, vertices_number, min_edge, max_edge, sign)
     print("Generating . . . ")
     second_time = time.perf_counter()
     elapsed_time = second_time - first_time
