+++ pyrgg/test.py
@@ -1185,6 +1185,74 @@ edge(27,14,7).
 26 19 28
 27 14 7
 <BLANKLINE>
+>>> random.seed(2)
+>>> dl_maker("testfile", 0, 200, 10, 0, 2, 0)
+7
+>>> file=open("testfile.dl","r")
+>>> print(file.read())
+dl
+format=edgelist1
+n=10
+data:
+4 3 -64
+5 4 148
+5 11 110
+6 7 -139
+6 8 -9
+8 8 -97
+9 9 143
+<BLANKLINE>
+>>> random.seed(4)
+>>> dl_maker("testfile2",0,50,30,0,4,0)
+41
+>>> file=open("testfile2.dl","r")
+>>> print(file.read())
+dl
+format=edgelist1
+n=30
+data:
+1 10 46
+2 16 5
+2 3 25
+2 18 -48
+4 17 -17
+5 27 16
+6 31 41
+7 26 -12
+7 6 -18
+8 3 -42
+8 13 11
+9 16 -5
+10 27 0
+10 30 -36
+10 23 -48
+10 17 26
+11 20 -27
+11 15 14
+11 10 -2
+14 21 -33
+14 18 -44
+14 11 43
+15 3 -12
+16 9 22
+16 14 -40
+16 18 20
+19 9 7
+20 30 18
+20 15 2
+21 23 18
+21 24 -1
+22 10 -9
+22 25 -39
+24 20 28
+24 10 16
+25 20 21
+25 19 23
+25 2 -10
+26 26 -18
+26 19 28
+27 14 7
+<BLANKLINE>
 >>> cov.stop()
 >>> cov.save()
 
