### 2017-08-30 3371e92
docs : minor edit in output image width
