### 2017-08-30 0814a16
docs : scrutinizer badge deleted
