### 2017-08-30 #331 @sepandhaghighi
Hi

Sorry for my late reply
I added more features to project (new file formats, update documents and [webpage](http://pyrgg.shaghighi.ir), code coverage and ...)

1-Statement of need : I added this section to Readme and Paper :

> pyrgg is an easy to use synthetic random graph generator written in python, that support variant graph file formats include DIMACS(.gr)	pyrgg has the ability to generate graphs in different sizes and designed to provide input files for test, benchmark and performance-analysis of graph processing frameworks, but it can be used in any other research program that include graphs with these formats.

2- Example Of Usage : I added [Example Of Usage](https://github.com/sepandhaghighi/pyrgg#example-of-usage) section to Readme

3- References are available in [paper.bib](https://github.com/sepandhaghighi/pyrgg/blob/master/paper/paper.bib)

Best Regards
Sepand Haghighi
