### 2017-09-06 #331 @hausen
@sepandhaghighi

1 - Statement of need: I can understand the text, but it would be better if you had someone proofread it.

2 - Example of usage: Can you please provide detailed instructions on how you produced the data for the bar graph in the "example of usage" section?

3 - References: ok
