### 2017-09-06 eaf4b41
docs : README and PAPER file updated
