+++ paper/paper.md
@@ -19,9 +19,8 @@ date: 18 July 2017
 
 # Summary
 
-Pyrgg is an easy to use synthetic random graph generator written in python, that support variant graph file formats include DIMACS.																										
-Pyrgg has the ability to generate graphs in different sizes and designed to provide input files for 
-test, benchmark and performance analysis of graph-processing frameworks, but it can be used in any other research program that include graphs with these formats.			
+Pyrgg is an easy-to-use synthetic random graph generator written in Python which supports various graph file formats including DIMACS .gr files.
+Pyrgg has the ability to generate graphs of different sizes and is designed to provide input files for broad range of applications, including but not limited to testing, benchmarking and performance-analysis of graph processing frameworks.		
 
 Graph Specifications :
 
