+++ README.md
@@ -18,10 +18,9 @@
 </div>				
 				
 ## Overview			
+pyrgg is an easy-to-use synthetic random graph generator written in Python which supports various graph file formats including <a href ="http://www.diag.uniroma1.it/challenge9/format.shtml">DIMACS .gr </a> files.
+pyrgg has the ability to generate graphs of different sizes and is designed to provide input files for broad range of applications, including but not limited to testing, benchmarking and performance-analysis of graph processing frameworks.
 
-pyrgg is an easy to use synthetic random graph generator written in python, that support variant graph file formats include <a href ="http://www.diag.uniroma1.it/challenge9/format.shtml">DIMACS(.gr)	</a> 									
-pyrgg has the ability to generate graphs in different sizes and designed to provide input files for 
-test, benchmark and performance-analysis of graph processing frameworks, but it can be used in any other research program that include graphs with these formats.		
 
 ## Installation		
 
