### 2017-09-06 07a5c4a
docs : README file example of usage updated
