### 2017-09-06 b668b5c
docs : minor edit in table alignment
