+++ README.md
@@ -153,7 +153,7 @@ or send an email to [sepand@qpage.ir](mailto:sepand@qpage.ir "sepand@qpage.ir").
 <img src="http://www.shaghighi.ir/pyrgg/images/performance.png">
 <p>Example. Performance comparison between Medusa and Basic implementation (RMAT&Random are synthetic)</p>
 <div align="center">
-<table style="border:1px solid black">
+<table style="border:1px solid black;" align="center">
 <tr>
 <th>Graph</th>
 <th>Vertices (10^6)</th>
@@ -226,7 +226,7 @@ or send an email to [sepand@qpage.ir](mailto:sepand@qpage.ir "sepand@qpage.ir").
 <p>Graphs Details</p>
 </div>
 <img src="http://www.shaghighi.ir/pyrgg/images/random.png">
-<p>Random Graph Generation</p>
+<p>Rand Graph Generation</p>
 </div>
 
 - Generate synthetic data for graph benchmark suite like [GAP](https://github.com/sbeamer/gapbs) 
