### 2017-09-06 8d55494
fix : Generating Print function relocated
