### 2017-09-06 208f494
docs: graph tables location modified
