### 2017-09-06 ed284f7
docs: graph tables location modified
