### 2017-09-06 78b96bd
feat : filesize function added
