### 2017-09-06 abb39d5
fix : minor edit in get_input function print formats
