### 2017-09-06 958def7
docs : README file modified
