### 2017-09-06 4c301d2
docs : README file modified
