### 2017-09-06 5fa25f5
docs : README file modified in example of usage section
