### 2017-09-06 #331 @sepandhaghighi
@hausen

1 - Statement of need :  I've modified the text -->

> Pyrgg is an easy-to-use synthetic random graph generator written in Python which supports various graph file formats including DIMACS .gr files.
Pyrgg has the ability to generate graphs of different sizes and is designed to provide input files for broad range of graph-based research applications, including but not limited to testing, benchmarking and performance-analysis of graph processing frameworks.

2 - Example of usage : I added  Table. 1. Characteristics of graphs & Fig. 2. Rand Graph Generation in [Example of Usage section](https://github.com/sepandhaghighi/pyrgg#example-of-usage) .
Fig. 1(bar graph) is about runtime comparison between Medusa and pure CUDA implementation of SSSP for different graphs.

Best Regards
Sepand Haghighi
