### 2017-09-11 #331 @danielskatz
@sepandhaghighi:

For statement of need, both the paper and the software documentation should state "what problems the software is designed to solve and who the target audience is?"  Think about this from the point-of-view of someone who is not an expert in your software.  How will they know if they should use this software?

For example of usage, the software documentation needs to "include examples of how to use the software (ideally to solve real-world analysis problems)."  Think again of a person who is not familiar with you software, and provide step-by-step examples of how they would use your software.
