### 2017-09-12 #331 @sepandhaghighi
@hausen , Bar graph data obtained from Medusa and pure CUDA implementation of SSSP for different graphs, and these steps are out of this paper scope(related to other frameworks). I added Fig.1 only to show this benchmark includes random graph.
