### 2017-09-12 #331 @sepandhaghighi
@danielskatz :
1. statement of need
> pyrgg is an easy-to-use synthetic random graph generator written in Python which supports various graph file formats including DIMACS .gr files. pyrgg has the ability to generate graphs of different sizes and is designed to provide input files for broad range of graph-based research applications, including but not limited to testing, benchmarking and performance-analysis of graph processing frameworks.

I think is clear, isn't ?

2. Example of usage
I added step-by-step examples of how generate graphs with different formats and also described each format separately but next steps (about graph analysis) are out of this paper scope (related to other frameworks).
