### 2017-09-12 #331 @danielskatz
Can you add to the paper and the documentation something that helps say "who the target audience is"?  What types of researchers might use this software?
