### 2017-09-12 #331 @danielskatz
Personally, I think the examples of usage part is fine, with the videos.  It's not exactly "(ideally to solve real-world analysis problems)" but this software doesn't solve analysis problems, so it's ok.
