### 2017-09-12 ef74af7
docs : target audience added to README and PAPER
