+++ paper/paper.md
@@ -20,7 +20,8 @@ date: 18 July 2017
 # Summary
 
 Pyrgg is an easy-to-use synthetic random graph generator written in Python which supports various graph file formats including DIMACS .gr files.
-Pyrgg has the ability to generate graphs of different sizes and is designed to provide input files for broad range of graph-based research applications, including but not limited to testing, benchmarking and performance-analysis of graph processing frameworks.		
+Pyrgg has the ability to generate graphs of different sizes and is designed to provide input files for broad range of graph-based research applications, including but not limited to testing, benchmarking and performance-analysis of graph processing frameworks.	
+Pyrgg target audiences are computer scientists who study graph algorithms and graph processing frameworks.	
 
 Graph Specifications :
 
