### 2017-09-12 #331 @danielskatz
Thanks @sepandhaghighi

@hausen - I think this is sufficient for the statements of need, and that the example usage is also ok, but I defer to you to make your recommendation.
