### 2017-09-16 #331 @hausen
Well, that's probably as good as it gets without a real-world example of usage.

My last two comments:

1) since it's "out of this paper scope" to show how the data for Fig 1 and Table 1 was processed, please remove them both from the README.md. IMHO, displaying this information -- without at least mentioning how you went from the output of the program to the data that gave birth to it -- only confuses a potential user of your program.

2) in the future, do update the repository with links to published research results that used your code

Once 1 is done, we're good to go!
