### 2017-09-16 3271839
docs : Fig.1 and Table.1 deleted
