### 2017-09-16 #331 @sepandhaghighi
@hausen

1. Removed

2. I will add links in the future.
