### 2017-09-17 #331 @arfon
@sepandhaghighi - the citations aren't being compiled in by Pandoc as you're not citing them in the body of the paper. Could you cite them directly please? (You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html))
