### 2017-09-17 263bc12
docs : paper file ref updated
