+++ paper/paper.bib
@@ -1,21 +1,15 @@
-@article{Zhong:2012:OMS:2370036.2145855,
- author = {Zhong, Jianlong and He, Bingsheng},
- title = {An Overview of Medusa: Simplified Graph Processing on GPUs},
- journal = {SIGPLAN Not.},
- issue_date = {August 2012},
- volume = {47},
- number = {8},
- month = feb,
- year = {2012},
- issn = {0362-1340},
- pages = {283--284},
- numpages = {2},
- url = {http://doi.acm.org/10.1145/2370036.2145855},
- doi = {10.1145/2370036.2145855},
- acmid = {2145855},
- publisher = {ACM},
- address = {New York, NY, USA},
- keywords = {GPGPU, GPU programming, graph processing, runtime framework},
+@article{Zhong2012,
+  doi = {10.1145/2370036.2145855},
+  url = {https://doi.org/10.1145/2370036.2145855},
+  year  = {2012},
+  month = {sep},
+  publisher = {Association for Computing Machinery ({ACM})},
+  volume = {47},
+  number = {8},
+  pages = {283},
+  author = {Jianlong Zhong and Bingsheng He},
+  title = {An overview of Medusa},
+  journal = {{ACM} {SIGPLAN} Notices}
 }
 
 @incollection{Chakrabarti2004,
